+++ base/lib/src/storage/task/request_task.dart
@@ -29,7 +29,7 @@ abstract class RequestTask<T> extends Task<T> {
       controller?.notifyStatusListeners(RequestTaskStatus.Request);
       options
         ..cancelToken = controller?.cancelToken
-        ..onSendProgress = controller?.notifyProgressListeners;
+        ..onSendProgress = onSendProgress;
 
       return options;
     }));
@@ -54,9 +54,9 @@ abstract class RequestTask<T> extends Task<T> {
   void postError(Object error) async {
     // 重试和冻结
     if (error is DioError) {
-      if (!canConnectToHost(error)) {
+      if (!_canConnectToHost(error)) {
         // host 连不上，判断是否 host 不可用造成的, 比如 tls error(没做还)
-        if (isHostUnavailable(error)) {
+        if (_isHostUnavailable(error)) {
           config.hostProvider.freezeHost(error.request.path);
         }
 
@@ -69,7 +69,7 @@ abstract class RequestTask<T> extends Task<T> {
       }
 
       // 能连上但是服务器不可用，比如 502
-      if (isHostUnavailable(error)) {
+      if (_isHostUnavailable(error)) {
         config.hostProvider.freezeHost(error.request.path);
 
         // 切换到其他 host
@@ -106,8 +106,13 @@ abstract class RequestTask<T> extends Task<T> {
     super.postError(error);
   }
 
+  // 自定义发送进度处理逻辑
+  void onSendProgress(int sent, int total) {
+    controller?.notifyProgressListeners(sent, total);
+  }
+
   // host 是否可以连接上
-  bool canConnectToHost(Object error) {
+  bool _canConnectToHost(Object error) {
     if (error is DioError) {
       if (error.type == DioErrorType.RESPONSE &&
           error.response.statusCode > 99) {
@@ -123,7 +128,7 @@ abstract class RequestTask<T> extends Task<T> {
   }
 
   // host 是否不可用
-  bool isHostUnavailable(Object error) {
+  bool _isHostUnavailable(Object error) {
     if (error is DioError) {
       if (error.type == DioErrorType.RESPONSE) {
         final statusCode = error.response.statusCode;
